rename sbt.* packages to be more consistent with the rest of the packages
